Add files via upload
The problem addressed in this project is the generation of handwritten digit images using a Generative Adversarial Network (GAN). GANs are a class of machine learning frameworks designed to generate new data samples that resemble a given dataset. The specific problem in this case is to create a model capable of generating images of handwritten digits (0-9) that are indistinguishable from real images in the MNIST dataset, a widely used benchmark dataset in the field of image processing and computer vision.